add transaction hash and timestamp to account subscription callbacks

+++ lib/services/v2.0/types.dart
@@ -15,7 +15,7 @@ class Event {
 
 class TransactionMetadata {
   String hash;
-  int timestamp;
+  BigInt timestamp;
 
   TransactionMetadata(this.hash, this.timestamp);
 }
